ng2: move common.ng2module to new file
