+++ generators/client-2/templates/src/main/webapp/app/components/_common.ng2module.ts
@@ -1,0 +1,19 @@
+import { XSRFStrategy, CookieXSRFStrategy } from '@angular/http';
+import { FormsModule } from '@angular/forms';
+import { BrowserModule } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+
+import { ProfileService } from './profiles/profile.service';
+
+@NgModule({
+    imports: [BrowserModule, FormsModule],
+    providers: [
+        ProfileService,
+        {
+            provide: XSRFStrategy,
+            useValue:  new CookieXSRFStrategy('CSRF-TOKEN', 'X-CSRF-TOKEN')
+        }
+    ],
+    declarations: []
+})
+export class <%=angularAppName%>CommonModule {}
